Valid swagger specification 2.0 implementation

+++ lib/Model/CreateList.php
@@ -64,7 +64,7 @@ class CreateList implements ArrayAccess
       */
     protected static $swaggerFormats = [
         'name' => null,
-        'folderId' => 'int32'
+        'folderId' => 'int64'
     ];
 
     public static function swaggerTypes()
