+++ lib/Model/GetExtendedListCampaignStats.php
@@ -63,7 +63,7 @@ class GetExtendedListCampaignStats implements ArrayAccess
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'campaignId' => 'int32',
+        'campaignId' => 'int64',
         'stats' => null
     ];
