+++ lib/Model/GetFolder.php
@@ -66,11 +66,11 @@ class GetFolder implements ArrayAccess
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'id' => 'int32',
+        'id' => 'int64',
         'name' => null,
-        'totalBlacklisted' => 'int32',
-        'totalSubscribers' => 'int32',
-        'uniqueSubscribers' => 'int32'
+        'totalBlacklisted' => 'int64',
+        'totalSubscribers' => 'int64',
+        'uniqueSubscribers' => 'int64'
     ];
 
     public static function swaggerTypes()
