+++ lib/Model/GetFolders.php
@@ -64,7 +64,7 @@ class GetFolders implements ArrayAccess
       */
     protected static $swaggerFormats = [
         'folders' => null,
-        'count' => 'int32'
+        'count' => 'int64'
     ];
 
     public static function swaggerTypes()
